2022/template/shortest_walk: Don’t count shortest paths if backtracked
If the visit function reported that the move is invalid then we
shouldn’t register it as a path to that point. This previously didn’t
matter because it was assumed the visit function was checking whether
points were visitable and the source point didn’t matter. If we want to
have a maze where points are visitable only from certain directions then
it is important to make a distinction about whether to backtrack because
the block is not visitable or because the goal has been reached.