2022/16: Add a generic const param to the walker for the num actors
The actors are used to represent the human and the elephants. Each
position and action is now an array representing what happens for each
actor. At each step we try all different combinations of actions. There
is now also an action StayStill to handle the case where one of the
actors can no longer do anything useful but the other one should still
move.