2022/16: Don’t revisit a valve if *any* actor was there
Instead of checking that the actor in question hasn’t already visited a
valve since the last valve opening, check whether *any* actor was there.
If an actor was there then it would should always have been more
efficient for them to open it at that point rather than letting another
actor return to it.