2022/16: Just iterate the permutations!
Instead of trying to walk the graph to find the solution, the graph is
only walked once for each start point up front to build up a lookup
table of distances between valves. Then for part 1, we try every
permutation of valves that have a flow and calculate the score for
visiting the valves in that order using the distance table. For part 2
we split the valves into every possible subset. Then we run the
algorithm from part 1 for each subset and combine it with the score of
running it on everything that isn’t in the subset.