2022/19: Try making a series of geode robots
Instead of trying all combinations until we use up all the minutes, it
now keeps trying until it makes a geode robot and then restarts the
search after that.

This doesn’t work because there are multiple ways to build a geode robot
with the same number of minutes and it doesn’t pick the best one.